Fix a kernel panic with sample mem_slab
It was caused by the mem_slab object (or any object) to be corrupted
when the thread was woken up by a timer expiration
but the end of  z_pend_current_on() function was not immediately processed (after the z_yield()).

We now make sure to remove expired threads from the waitqueue it was pending on if any.

Add readm to mem_slab sample.